// src/app/enums/CommandType.ts
export enum CommandType {

	NORTH = 'north',
	EAST = 'east',
	SOUTH = 'south',
	WEST = 'west',
	LOOK_NORTH = 'lookNorth',
	LOOK_EAST = 'lookEast',
	LOOK_SOUTH = 'lookSouth',
	LOOK_WEST = 'lookWest',
	FAIL_FORWARD = 'failForward',
	FAIL_BACKWARD = 'failBackward',
	TURN_LEFT = 'turnLeft',
	TURN_RIGHT = 'turnRight',
	FINISH = 'finish'

}